+++ my-vue-app/src/Components/Card.jsx
@@ -1,7 +1,10 @@
+// import { useSelector } from "react-redux";
 // import { Link } from "react-router-dom";
 
 // export default function Card(props) {
 //   // eslint-disable-next-line react/prop-types
+//   const cart = useSelector((state) => state.cartByUserId)  // cartByUserId.qunatity = 2
+//   // console.log(cartUser)
 //   let { id, name, description, price, imageUrl } = props;
 //   return (
 //     <div className="flex flex-col p-4 bg-gradient-to-br from-white via-gray-100 to-gray-300 mb-5 rounded-sm max-w-sm border border-gray-300">
@@ -26,23 +29,26 @@
 //     </div>
 //   );
 // }
+
+import { useDispatch } from "react-redux";
 import { Link } from "react-router-dom";
-import { useState } from "react";
+import { deleteItem } from "../Redux/actions";
 
 export default function Card(props) {
-  const { id, name, description, price, imageUrl } = props;
-  const [quantity, setQuantity] = useState(1);
-
-  const handleIncrement = () => {
-    setQuantity(quantity + 1);
-  };
-
-  const handleDecrement = () => {
-    if (quantity > 1) {
-      setQuantity(quantity - 1);
-    }
-  };
-
+  const dispatch = useDispatch()
+  const idLocal = localStorage.getItem("id");
+  let { id, name, description, price, imageUrl } = props;
+  const cart = [{productId: 1,quantity: 3}, {quantity: 2}, {quantity: 20}, {quantity: 10}, {quantity: 1}];
+  // const array2 = cart.filter(e => e.quantity > 1);
+  
+  // console.log(array2);
+  const itemInCart = cart.find(item => item.productId === id);
+  const itemCount = itemInCart ? itemInCart.quantity : 1;
+  const data = {userId: parseInt(idLocal), itemId:id}
+  console.log(data)
+  const handleClick = () => {
+    dispatch(deleteItem(data))
+  }
   return (
     <div className="flex flex-col p-4 bg-gradient-to-br from-white via-gray-100 to-gray-300 mb-5 rounded-sm max-w-sm border border-gray-300">
       <figure className="w-full flex items-center justify-center">
@@ -50,7 +56,7 @@ export default function Card(props) {
       </figure>
       <br />
       <h3 className="px-4 py-2 text-base text-white font-medium bg-[#0061FB] p-1 text-center rounded-sm">
-        {name}
+        {name} 
       </h3>
       <p className="px-4 py-2 text-base font-medium text-[#484848]">
         {description}
@@ -58,32 +64,17 @@ export default function Card(props) {
       <p className="px-4 py-2 text-base font-medium text-[#484848]">
         $ {price}
       </p>
-      <div className="flex justify-center">
-        <button
-          className="flex justify-center bg-nintendo text-white font-medium px-4 py-2 rounded-sm"
-          onClick={handleDecrement}
-        >
-          -
-        </button>
-        <input
-          type="number"
-          value={quantity}
-          className="w-12 text-center bg-white border border-gray-300"
-          readOnly
-        />
-        <button
-          className="flex justify-center bg-nintendo text-white font-medium px-4 py-2 rounded-sm"
-          onClick={handleIncrement}
-        >
-          +
-        </button>
-      </div>
       <Link className="flex justify-center" to={`/product/${id}`}>
         <h2 className="flex justify-center bg-nintendo text-white font-medium px-4 py-2 rounded-sm">
           <p>Detalle</p>
         </h2>
       </Link>
+      <br />
+      <button className="flex justify-center bg-nintendo text-white font-medium px-4 py-2 rounded-sm"
+      onClick={handleClick}>Quitar</button>
+      <h1>{itemCount > 1 && <span>Cantidad: {itemCount}</span>}</h1>
     </div>
   );
 }
+
